--- Assets/Scripts/Button/LoginButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0eec1603e55a5a448ff27d1f30474d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
