+++ Assets/Scripts/Entities/TopDownAimRotation.cs
@@ -19,11 +19,17 @@ public class TopDownAimRotation : MonoBehaviour
     void Start()
     {
         _controller.OnLookEvent += OnAim;
+        _controller.OnMoveEvent += OnMoveRotation;
     }
 
     private void OnAim(Vector2 newAimDirection)
     {
         RotateArm(newAimDirection);
+    }
+
+    private void OnMoveRotation(Vector2 newMoveDirection)
+    {
+        RotateCharacter(newMoveDirection);
     }
 
     // Atan2 = Atan의 값을 계산해서 세타값을 구한다 ( 백터의 각도를 구하는 것)
@@ -35,11 +41,12 @@ public class TopDownAimRotation : MonoBehaviour
         //armRenderer.flipY = Mathf.Abs(rotZ) > 90f;
         //characterRenderer.flipX = armRenderer.flipY;
         // armPivot.rotation = Quaternion.Euler(0, 0, rotZ);
+
     }
 
-  
-    void Update()
+    private void RotateCharacter(Vector2 direction)
     {
-        
+        float rotZ = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
+        characterRenderer.flipX = Mathf.Abs(rotZ) > 90f;
     }
 }
