--- Assets/Scripts/Player/PlayerEnterName.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f63e2025afabb6c4f9a26ed74d390b6f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
