+++ Assets/Scripts/Player/PlayerName.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2546ab71d26f92943b38defa0c0057f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
